Layer API Refactoring
