+++ tensorlayer/layers/spatial_transformer.py
@@ -12,7 +12,7 @@ from tensorlayer.layers.core import Layer
 from tensorlayer.layers.core import LayersConfig
 from tensorlayer.layers.core import TF_GRAPHKEYS_VARIABLES
 
-from tensorlayer.layers.core import flatten_reshape
+from tensorlayer.layers.utils import flatten_reshape
 
 from tensorlayer import tl_logging as logging
